Rework the core logic of provider in Rust (#872)
* Migrate to new RPC impl and fix most of warnings

* Merge some method matching branches

* Merge foo/on_init method matching

* Make on_create() fallible

* Clean up write_response in dumb_jump/on_move impl

* Hide RpcClient behind Vim

* Unify api in Vim

* OnMoveHandler always expects curline

* Remove msg arg from OnMove event

* Fix the unused warnings

* Fix file preview

* Fix display_getcurline()

* Rework recent_files provider Rust impl

* Bare Create

* Bare OnTyped event

* Migrate call to notify: on_typed

* Migrate call to notify: on_move

* Delete deprecated_runner.rs

* Do not break users using older version

* Merge session_client.rs

* Clean up on_create impl

* Handle context query for dumb_jump

* Initialize ext_map in initialize_global_env

* Remove print_on_typed and add Value varint to Params

* Remove debugging print

* Simplify grep2 init

* Check if curline is empty

* Revert single Value in params

* Store matcher in Context

* Exit event_loop when terminated

* Store current_results in DefaultProvider::on_typed and simplify the method in Vim request

* Merge session/manager.rs

* Remove Call.session_id()

* Extract vim.set_var()

* Refactor PreviewKind

* .

* Add debounce to provider attribute

* Refactor ProviderSource

* Set forerunner_temp_file on_create

* Refactor create_new_source()

* Fix init_display_win()

* Set ProviderSource::Command

* Refactor job.rs

* Small clean up

* Refactor provider.rs and session.rs

* Remove SessionState

* Use entry API

* win_execute() has been patched to neovim

* Refactor ProviderSource

* Tweak the debounce delay according to the source scale

* Change provider_source to RwLock

* Specify icon in provider

* Add s:recent_files.icon

* Initialize display_winwidth on Rust side

* Add on_move debounce

* Make display_curline() just one request

* Retrieve &runtimepath on demand

* Simplify api.vim

* Make params in clap#client#notify() optional

* Parse grep icon

* Reimplement help_tags on the Rust side

* Impl grep forerunner on Rust backend

* Remove impl/on_typed.vim

* Remove job/forerunner.vim

* Extract a StdioProgressor wrapper type

* Make par_dyn_run generic over progressor

* Introduce VimProgressor and support par_dyn_run on Rust

* Fix on_typed Rust impl and migrate to deprecated_method

Basically use tokio::task::spawn_blocking when necessary.

* Abort the potential on_typed blocking task in handle_terminate()

* Extract par_dyn_run_inprocess

* Specify detached for Exec in par_dyn_run_inprocess

* Clean up printer and make args of vim.exec() a List

* [grep2]Only reuse the cache directly when it's huge

* Make Vim a field of SessionContext

* Rename SessionContext to ProviderContext

* Simplify clap#client#notify_on_init()

* Try delegate on_typed grep2 to Rust

* Make default target to maple in Makefile

* Reorg stdio_server

* Display <Empty file> when the preview content is empty

* Initialize indicator

* Move ProviderContext to provider/mod.rs

* Reimplement filer in Rust

* Refactor ProviderEvent::KeyTyped

* Clean up filer

* Update rust version to make use of [default]

* More debug info

* Refactor read_preview_lines()

* Add buffers.on_move_async impl

* Ensure no multiple par_dyn_run_inprocess spawned

* Extract ProviderEnvironment and rename to PreviewTarget

* Introduce Preview struct

* Refactor ProviderEvent

* Indicate the cache being using in UI

* Add stopped_signal to VimProgressor

* Remove GlobalEnv

* Fix merge

* Fix vint

* Refactor grep module

* Extract try_into_clap_item()

* [neovim] Add <Alt-u> for the back action in the normal mode

Close #888

* Rename Source to SequentialSource

* Rename to `filter_sequential()`

* Reorg filter crate

* Rename to `ParallelSource` and `to_clap_item()`

* Reorder imports

* Refactor grep_impl.rs

* Move grep.rs to grep/mod.rs

* [ci] Update clippy check

* Fix CI

* Fix clippy

Fix clippy

Refactor BufferInitializationResult

More clippy fixes

* Initial ripgrep module

* FMT

* Fix test

* clippy

* clippy

* clippy

* MatcherBuilder::new()

* InverseMatcherWithRecord

* Introduce ExecutableSearcher in searcher_impl.rs

* match_jump_line()

* WordRegexSearcher, LanguageRegexSearcher

* executable_searcher.rs

* SearcherMessage

* Searcher module

* Update searcher progress

* Prototype of `grep` on libgrep

* Clean up prototype

* Split out GrepProvider

* Clean up GrepProvider

* generic_provider

* StoppableSearcher

* Fix CI

* Refactor ProgressUpdate

* Nits

* Skip grep initialization

* BlinesProvider

* on_terminate()

* Fix highlights in FileResult

* trim line

* Fix matcher test

* .

* Fix match indices in line

* PreviewImpl

* bare_call() and bare_exec()

* Nits

* Restore upgrade test

* Update README.md

* blines: clean screen on empty query

* Fix vint

* FilesProvider

* files: take care of +name-only

* Minor refactorings

* Merge ripgrep/util.rs into ripgrep/stats.rs

* replace_range()

* Update CHANGELOG.md

+++ crates/maple_cli/src/command/grep/forerunner.rs
@@ -1,14 +1,11 @@
-use std::path::PathBuf;
-
-use anyhow::Result;
-use clap::Parser;
-
-use utility::is_git_repo;
-
 use super::{rg_command, rg_shell_command};
 use crate::app::Params;
 use crate::process::CacheableCommand;
 use crate::utils::{send_response_from_cache, SendResponse};
+use anyhow::Result;
+use clap::Parser;
+use std::path::PathBuf;
+use utility::is_git_repo;
 
 #[derive(Parser, Debug, Clone)]
 pub struct RipGrepForerunner {
@@ -60,7 +57,7 @@ impl RipGrepForerunner {
                     if digest.total > 100000 {
                         send_response_from_cache(
                             &digest.cached_path,
-                            digest.total as usize,
+                            digest.total,
                             SendResponse::Json,
                             icon,
                         );
